Build image on an instance from a local Dockerfile

Uploads a folder with a Dockerfile and supporting files to an instance and builds it

docker_build(host = localhost, dockerfolder, new_image,
  folder = "buildimage", wait = FALSE, ...)

Arguments

host

A host object.

dockerfolder

Local location of build directory including valid Dockerfile

new_image

Name of the new image

folder

Where on host to build dockerfile

wait

Whether to block R console until finished build

...

Other arguments passed to the SSH command for the host

Value

A table of active images on the instance

Details

Dockerfiles are best practice when creating your own docker images, rather than logging into a Docker container, making changes and committing.

See also

https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/

An example Dockerfile for https://hub.docker.com/r/rocker/ropensci/~/dockerfile/

General R Docker images found at https://github.com/rocker-org

Examples

## Not run: ------------------------------------
# docker_build(localhost, "/home/stuff/dockerfolder" ,"new_image", wait = TRUE)
# docker_run(localhost, "new_image")
## ---------------------------------------------