Run a docker command on a host.

Run a docker command on a host.

docker_cmd(host, cmd = NULL, args = NULL, docker_opts = NULL,
  capture_text = FALSE, ...)

Arguments

host

A host object.

cmd

A docker command, such as "run" or "ps"

args

Arguments to pass to the docker command

docker_opts

Options to docker. These are things that come before the docker command, when run on the command line.

capture_text

If FALSE (the default), return the host object. This is useful for chaining functions. If TRUE, capture the text output from both stdout and stderr, and return that. Note that TRUE may not be available on all types of hosts.

...

Other arguments passed to the SSH command for the host

Examples

## Not run: ------------------------------------
# docker_cmd(localhost, "ps", "-a")
## ---------------------------------------------