Docker S3 method for use with harbor package

Docker S3 method for use with harbor package

# S3 method for gce_instance
docker_cmd(host, cmd = NULL, args = NULL,
  docker_opts = NULL, capture_text = FALSE, ...)

Arguments

host

The GCE instance

cmd

The command to pass to docker

args

arguments to the command

docker_opts

options for docker

capture_text

whether to return the output

...

other arguments passed to gce_ssh

Details

Instances launched in the google-containers image family automatically add your user to the docker group, but for others you will need to run sudo usermod -a -G docker ${USER} and log out and back in.