Build a dockerfile on a remote VM

Instead of building locally or via SSH, launch a dedicated VM

gce_docker_build(dockerfile, build_name, ...)

Arguments

dockerfile

The Dockerfile to build from

build_name

The name of the Docker image to build

...

Other parameters passed to gce_vm_template

Details

This avoids using SSH by creating a dedicated VM to build your image, and upload it to the Google Container Registry.

It uses the builder template available via gce_vm_template

If not otherwise specified as arguments, the name of the VM will be dockerbuild and the predefined_type will be n1-standard-1

TODO: After uploading it will then delete the build VM

See also

Other container registry functions: gce_list_registry, gce_pull_registry, gce_push_registry, gce_tag_container

Examples

## Not run: ------------------------------------
# 
# ## A dockerfile that comes with the package to install googleAuthR packages
# my_dockerfile <- system.file("dockerfiles",
#                              "googleAuthR-verse", 
#                              "Dockerfile", 
#                              package = "googleComputeEngineR")
# ## build the dockerfile on remote VM                             
# gce_docker_build(dockerfile = my_dockerfile, 
#                  build_name = "test_docker_image")
# 
## ---------------------------------------------