Install R packages onto an instance's stopped docker image

Install R packages onto an instance's stopped docker image

gce_future_install_packages(instance, docker_image, cran_packages = NULL,
  github_packages = NULL)

Arguments

instance

The instance running the container

docker_image

A docker image to install packages within.

cran_packages

A character vector of CRAN packages to be installed

github_packages

A character vector of devtools packages to be installed

Value

TRUE if successful

Details

See the images on the instance via harbor::docker_cmd(instance, "images")

If using devtools github, will look for an auth token via devtools::github_pat(). This is an environment variable called GITHUB_PAT

Will start a container, install packages and then commit the container to an image of the same name via docker commit -m "installed packages via gceR"