Returns the specified Zone resource. Get a list of available zones by making a list() request.

Returns the specified Zone resource. Get a list of available zones by making a list() request.

gce_get_zone(project, zone)

Arguments

project

Project ID for this request

zone

Name of the zone resource to return

Details

Authentication scopes used by this function are:

https://www.googleapis.com/auth/cloud-platform

https://www.googleapis.com/auth/compute

https://www.googleapis.com/auth/compute.readonly

See also

https://developers.google.com/compute/docs/reference/latest/