Retrieves the list of instances contained within the specified zone.

Retrieves the list of instances contained within the specified zone.

gce_list_instances(filter = NULL, maxResults = NULL, pageToken = NULL,
  project = gce_get_global_project(), zone = gce_get_global_zone())

Arguments

filter

Sets a filter expression for filtering listed resources, in the form filter=expression

maxResults

The maximum number of results per page that should be returned

pageToken

Specifies a page token to use

project

Project ID for this request

zone

The name of the zone for this request

Details

Authentication scopes used by this function are:

https://www.googleapis.com/auth/cloud-platform

https://www.googleapis.com/auth/compute

https://www.googleapis.com/auth/compute.readonly

To filter you need a single string in the form field_name eq|ne string e.g. gce_list_instances("status eq RUNNING") where eq is 'equals' and ne is 'not-equals'.

See also

https://developers.google.com/compute/docs/reference/latest/