Load a previously saved private Google Container

Load a previously saved private Google Container

gce_pull_registry(instance, container_name, container_url = "gcr.io",
  pull_only = FALSE, project = gce_get_global_project(), ...)

Arguments

instance
The VM to run within

container_name
The name of the saved container

container_url
The URL of where the container was saved

pull_only
If TRUE, will not run the container, only pull to the VM

project
Project ID for this request, default as set by gce_get_global_project

...
Other arguments passed to docker_run or docker_pull

After starting a VM, you can load the container again using this command.

For Shiny based containers, pass "-p 80:3838" to run it at the IP URL

For RStudio based containers, pass "-p 80:8787" to run it at the IP URL

Value

The instance

See also