Return a container tag for Google Container Registry

Return a container tag for Google Container Registry

gce_tag_container(container_name, project = gce_get_global_project(),
  container_url = "gcr.io")

Arguments

container_name
A running docker container. Can't be set if image_name is too.

project
Project ID for this request, default as set by gce_get_global_project

This will only work on the Google Container optimised containers of image_family google_containers. Otherwise you will need to get a container authentication yourself (for now)

It will start the push but it may take a long time to finish, especially the first time, this function will return whilst waiting but don't turn off the VM until its finished.

container_url
The URL of where to save container

Value

A tag for use in Google Container Registry

See also

Other container registry functions: gce_list_registry, gce_pull_registry, gce_push_registry