Create a template container VM

This lets you specify templates for the VM you wnat to launch It passes the template on to gce_vm_container

gce_vm_template(template = c("rstudio", "shiny", "opencpu", "r-base",
  "example", "rstudio-hadleyverse"), username = NULL, password = NULL,
  image_family = "gci-stable", name, ...)

Arguments

template

The template available

username

username if needed (RStudio)

password

password if needed (RStudio)

image_family

An image-family. It must come from the google-containers family.

name

The name of the resource, provided by the client when initially creating the resource

...

Other arguments passed to gce_vm_create

Value

The VM object

Details

Templates available are:

rstudio An RStudio server docker image

shiny A Shiny docker image

opencpu An OpenCPU docker image

r_base Latest version of R stable

example A non-R test container running busybox

Examples

## Not run: ------------------------------------
# 
#  library(googleComputeEngineR)
#  library(harbor)
#  
#  ## make instance using R-base
#  vm <- gce_vm_template("r-base", predefined_type = "f1-micro", name = "rbase")
#  
#  ## run an R function on the instance within the R-base docker image
#  docker_run(vm, "rocker/r-base", c("Rscript", "-e", "1+1"), user = "mark")
#  #> [1] 2
#  
# 
## ---------------------------------------------