Add Shiny app to a Shiny template instance

Add a local shiny app to a running Shiny VM installed via gce_vm_template

gce_shiny_addapp(instance, app_image, dockerfolder = NULL)

Arguments

instance
The instance running Shiny

app_image
The name of the Docker image to create or use existing from Google Container Registry. Must be numbers, dashes or lowercase letters only.

dockerfolder
The folder location containing the Dockerfile and app dependencies

Value

The instance

Details

To deploy a Shiny app, you first need to construct a Dockerfile which load the R packages and dependencies, as well as copying over the Shiny app in the same folder.

This function will take the Dockerfile, build it into a Docker image and upload it to Google Container Registry for use later.

If already created, then the function will download the app_image from Google Container Registry and start it on the instance provided.

Any existing Shiny Docker containers are stopped and removed, so if you want multiple apps put them in the same Dockerfile.

See also

The vignette entry called Shiny App has examples and a walk through.

Examples

An example \code{Dockerfile} is show below.  
This file in the same folder as your shiny app, which consists of a \code{ui.R} and \code{server.R} in a shiny subfolder.  
This is copied into the Dockerfile in the last line.  
Change the name of the subfolder to have that name appear in the final URL of the Shinyapp. 

not_run({
  FROM rocker/shiny
  MAINTAINER Mark Edmondson (r@sunholo.com)
  
  # install R package dependencies
  RUN apt-get update && apt-get install -y \
          libssl-dev \
          ## clean up
          && apt-get clean \ 
          && rm -rf /var/lib/apt/lists/ \ 
          && rm -rf /tmp/downloaded_packages/ /tmp/*.rds
          
  ## Install packages from CRAN
  RUN install2.r --error \ 
       -r 'http://cran.rstudio.com' \
       googleAuthR \
  ## install Github packages
  ## clean up
  && rm -rf /tmp/downloaded_packages/ /tmp/*.rds
  
  ## assume shiny app is in build folder /shiny
  COPY ./shiny/ /srv/shiny-server/shiny/
  
})

This is then run using the R commands below:

not_run({
  
  ## create shiny VM template
  vm <- gce_vm("shiny-test",
               template = "shiny",
               predefined_type = "n1-standard-1")
  
  ## add your SSH keys do docker commands work
  vm <- vm_ssh_setup(vm)
  
  ## a demo Shiny app from googleAuthR
  app_dir <- system.file("dockerfiles","shiny-googleAuthRdemo", package = "googleComputeEngineR")
  
  ## build Dockerfile on VM, and upload created Dockerimage to Container Registry
  gce_shiny_addapp(vm, app_image = "gceshinydemo", dockerfolder = app_dir)
  
  ## if you want to deploy the same Shiny app to a different VM, use the saved Container Registry 
  gce_shiny_addapp(vm2, app_image = "gceshinydemo")
  
})

#> Error: <text>:2:4: unexpected symbol
#> 1: 
#> 2: An example
#>       ^