Create a template container VM

This lets you specify templates for the VM you want to launch It passes the template on to gce_vm_container

gce_vm_template(template = c("rstudio", "shiny", "opencpu", "r-base",
  "dynamic", "rstudio-gpu", "rstudio-shiny"), username = NULL,
  password = NULL, dynamic_image = NULL, image_family = "cos-stable",
  wait = TRUE, ...)

Arguments

template
The template available

username
username if needed (RStudio)

password
password if needed (RStudio)

dynamic_image
Supply an alternative to the default Docker image for the template

image_family
An image-family. It must come from the cos-cloud family.

wait
Whether to wait for the VM to launch before returning. Default TRUE.

...
Arguments passed on to gce_vm_container

file

file location of a valid cloud-init or shell_script file. One of file or cloud_init or shell_script must be supplied

cloud_init

contents of a cloud-init file, for example read via readChar(file, nchars = 32768)

shell_script

contents of a shell_script file, for example read via readChar(file, nchars = 32768)

image_family

An image-family. It must come from the image_project family.

image_project

An image-project, where the image-family resides.

Value

The VM object, or the VM startup operation if wait=FALSE

Details

Templates available are:

rstudio An RStudio server docker image with tidyverse and devtools

rstudio-gpu An RStudio server with popular R machine learning libraries and GPU driver. Will launch a GPU enabled VM.

rstudio-shiny An RStudio server with Shiny also installed, proxied to /shiny

shiny A Shiny docker image

opencpu An OpenCPU docker image

r_base Latest version of R stable

dynamic Supply your own docker image within dynamic_image

For dynamic templates you will need to launch the docker image with any ports you want opened, other settings etc. via docker_run.

Use dynamic_image to override the default rocker images e.g. rocker/shiny for shiny, etc.

Examples

# NOT RUN {
 library(googleComputeEngineR)

 ## make instance using R-base
 vm <- gce_vm_template("r-base", predefined_type = "f1-micro", name = "rbase")

 ## run an R function on the instance within the R-base docker image
 docker_run(vm, "rocker/r-base", c("Rscript", "-e", "1+1"), user = "mark")
 #> [1] 2

# }